+++ src/inspect.rs
@@ -8,25 +8,36 @@ use biscuit_auth::{
 use chrono::offset::Utc;
 use serde::Serialize;
 use serde_json::json;
+use std::path::PathBuf;
 use std::{fmt::Display, fs};
-use std::{path::PathBuf, time::Duration};
 
 use crate::cli::*;
 use crate::errors::CliError::*;
 use crate::input::*;
 
-#[derive(Serialize)]
+#[derive(Serialize, Debug)]
 struct TokenBlock {
+    version: u32,
     code: String,
     external_key: Option<String>,
     revocation_id: String,
 }
 
-#[derive(Serialize)]
+#[derive(Serialize, Debug)]
 struct TokenDescription {
     sealed: bool,
     root_key_id: Option<u32>,
     blocks: Vec<TokenBlock>,
+}
+
+fn get_version_string(v: u32) -> String {
+    match v {
+        3 => "v3.0".to_string(),
+        4 => "v3.1".to_string(),
+        5 => "v3.2".to_string(),
+        6 => "v3.3".to_string(),
+        _ => "(unsupported version)".to_string(),
+    }
 }
 
 impl Display for TokenDescription {
@@ -54,7 +65,7 @@ impl Display for TokenDescription {
                 writeln!(f, "Block n°{}:", i)?;
             }
 
-            writeln!(f, "== Datalog ==")?;
+            writeln!(f, "== Datalog {} ==", get_version_string(block.version))?;
             writeln!(f, "{}", block.code)?;
 
             writeln!(f, "== Revocation id ==")?;
@@ -65,7 +76,7 @@ impl Display for TokenDescription {
     }
 }
 
-#[derive(Copy, Clone, Serialize)]
+#[derive(Copy, Clone, Serialize, Debug)]
 #[serde(untagged)]
 enum RResult<A, E> {
     Ok(A),
@@ -82,6 +93,13 @@ impl<A, E> From<std::result::Result<A, E>> for RResult<A, E> {
 }
 
 impl<A, E> RResult<A, E> {
+    pub fn is_err(&self) -> bool {
+        match self {
+            Self::Ok(_) => false,
+            Self::Err { .. } => true,
+        }
+    }
+
     pub fn into_result(self) -> std::result::Result<A, E> {
         match self {
             Self::Ok(a) => Ok(a),
@@ -90,20 +108,29 @@ impl<A, E> RResult<A, E> {
     }
 }
 
-#[derive(Serialize)]
+#[derive(Serialize, Debug)]
 struct QueryResult {
     query: String,
     query_all: bool,
     facts: RResult<Vec<String>, Token>,
+    elapsed_micros: u128,
 }
 
 impl Display for QueryResult {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         writeln!(f)?;
         if self.query_all {
-            writeln!(f, "🔎 Running query on all facts: {}", &self.query)?;
+            writeln!(
+                f,
+                "🔎 Running query on all facts: {} ({}μs)",
+                &self.query, self.elapsed_micros
+            )?;
         } else {
-            writeln!(f, "🔎 Running query: {}", &self.query)?;
+            writeln!(
+                f,
+                "🔎 Running query: {} ({}μs)",
+                &self.query, self.elapsed_micros
+            )?;
         }
         match &self.facts.clone().into_result() {
             Ok(facts) => {
@@ -123,12 +150,12 @@ impl Display for QueryResult {
     }
 }
 
-#[derive(Serialize)]
+#[derive(Serialize, Debug)]
 struct AuthResult {
     policies: Vec<String>,
     result: RResult<(usize, String), Token>,
     iterations: u64,
-    elapsed: Duration,
+    elapsed_micros: Option<u128>,
 }
 
 impl Display for AuthResult {
@@ -138,7 +165,9 @@ impl Display for AuthResult {
                 writeln!(
                     f,
                     "✅ Authorizer check succeeded 🛡️ ({}μs, {} iterations)",
-                    self.elapsed.as_micros(),
+                    self.elapsed_micros
+                        .map(|e| e.to_string())
+                        .unwrap_or("n/a".to_string()),
                     self.iterations,
                 )?;
                 writeln!(f, "Matched allow policy: {}", policy)
@@ -147,7 +176,9 @@ impl Display for AuthResult {
                 writeln!(
                     f,
                     "❌ Authorizer check failed 🛡️ ({}μs, {} iterations)",
-                    self.elapsed.as_micros(),
+                    self.elapsed_micros
+                        .map(|e| e.to_string())
+                        .unwrap_or("n/a".to_string()),
                     self.iterations,
                 )?;
                 match e {
@@ -160,7 +191,7 @@ impl Display for AuthResult {
     }
 }
 
-#[derive(Serialize)]
+#[derive(Serialize, Debug)]
 pub struct InspectionResults {
     token: TokenDescription,
     signatures_check: Option<bool>,
@@ -216,54 +247,80 @@ impl InspectionResults {
 struct SnapshotDescription {
     code: String,
     iterations: u64,
-    elapsed_micros: u128,
+    elapsed_micros: Option<u128>,
+    already_evaluated: bool,
 }
 
 impl Display for SnapshotDescription {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         writeln!(f, "{}", self.code)?;
 
-        writeln!(
-            f,
-            "⏱️ Execution time: {}μs ({} iterations)",
-            self.elapsed_micros, self.iterations
-        )
+        if let Some(elapsed_micros) = self.elapsed_micros {
+            writeln!(
+                f,
+                "⏱️ Execution time: {}μs ({} iterations)",
+                elapsed_micros, self.iterations
+            )
+        } else {
+            writeln!(f, "⏱️ Execution time: not evaluated yet")
+        }
     }
 }
 
 #[derive(Serialize)]
 pub struct SnapshotInspectionResults {
-    snapshot: SnapshotDescription,
-    auth: Option<AuthResult>,
-    query: Option<QueryResult>,
+    contents: SnapshotDescription,
+    evaluation: RResult<SnapshotEvaluationResults, Token>,
 }
 
 impl Display for SnapshotInspectionResults {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        self.snapshot.fmt(f)?;
+        self.contents.fmt(f)?;
 
-        match &self.auth {
-            None => writeln!(f, "🙈 Datalog check skipped 🛡️")?,
-            Some(auth_result) => auth_result.fmt(f)?,
-        }
-
-        match &self.query {
-            None => Ok(()),
-            Some(query_result) => query_result.fmt(f),
+        match &self.evaluation {
+            RResult::Ok(r) => r.fmt(f),
+            RResult::Err { error } => writeln!(f, "❌ Evaluation failed: {}", error),
         }
     }
 }
 
 impl SnapshotInspectionResults {
     pub fn ensure_success(&self) -> Result<()> {
-        if let Some(ref auth) = self.auth {
-            if auth.result.clone().into_result().is_err() {
-                Err(AuthorizationFailed)?;
-            }
+        match self.evaluation {
+            RResult::Ok(ref eval_result) => eval_result.ensure_success(),
+            RResult::Err { .. } => Err(EvaluationFailed)?,
+        }
+    }
+}
+
+#[derive(Serialize, Debug)]
+pub struct SnapshotEvaluationResults {
+    iterations: u64,
+    auth: AuthResult,
+    query: Option<QueryResult>,
+}
+
+impl Display for SnapshotEvaluationResults {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        self.auth.fmt(f)?;
+
+        match &self.query {
+            None => Ok(()),
+            Some(query_result) => query_result.fmt(f),
+        }?;
+
+        Ok(())
+    }
+}
+
+impl SnapshotEvaluationResults {
+    pub fn ensure_success(&self) -> Result<()> {
+        if self.auth.result.is_err() {
+            Err(AuthorizationFailed)?;
         }
 
-        if let Some(ref query) = self.query {
-            if query.facts.clone().into_result().is_err() {
+        if let Some(ref query_result) = self.query {
+            if query_result.facts.is_err() {
                 Err(QueryFailed)?;
             }
         }
@@ -277,7 +334,7 @@ fn handle_query(
     query_all: bool,
     all_params: &[Param],
     authorizer: &mut Authorizer,
-) -> Result<QueryResult> {
+) -> std::result::Result<QueryResult, Token> {
     let mut rule = query.clone();
 
     for p in all_params {
@@ -303,6 +360,10 @@ fn handle_query(
         facts: facts
             .map(|fs| fs.iter().map(|f| f.to_string()).collect::<Vec<_>>())
             .into(),
+        elapsed_micros: authorizer
+            .execution_time()
+            .expect("Missing execution time")
+            .as_micros(),
     })
 }
 
@@ -362,17 +423,41 @@ pub fn handle_inspect_inner(inspect: &Inspect) -> Result<InspectionResults> {
         &inspect.authorization_args.authorize_interactive,
         &inspect.authorization_args.authorize_with,
         &inspect.authorization_args.authorize_with_file,
+        &inspect.authorization_args.authorize_with_snapshot,
+        &inspect.authorization_args.authorize_with_snapshot_file,
     ) {
-        (false, None, None) => None,
-        (true, None, None) => Some(DatalogInput::FromEditor),
-        (false, Some(str), None) => Some(DatalogInput::DatalogString(str.to_owned())),
-        (false, None, Some(path)) => Some(DatalogInput::FromFile(path.to_path_buf())),
+        (false, None, None, None, None) => None,
+        (true, None, None, None, None) => Some(AuthorizerInput::FromDatalog(
+            DatalogInput::FromEditor,
+            inspect.param_arg.param.clone(),
+        )),
+        (false, Some(str), None, None, None) => Some(AuthorizerInput::FromDatalog(
+            DatalogInput::DatalogString(str.to_owned()),
+            inspect.param_arg.param.clone(),
+        )),
+        (false, None, Some(path), None, None) => Some(AuthorizerInput::FromDatalog(
+            DatalogInput::FromFile(path.to_path_buf()),
+            inspect.param_arg.param.clone(),
+        )),
+        (false, None, None, Some(str), None) => Some(AuthorizerInput::FromSnapshot(
+            SnapshotInput::FromString(str.to_owned()),
+        )),
+        (false, None, None, None, Some(path)) => {
+            Some(AuthorizerInput::FromSnapshot(SnapshotInput::FromFile(
+                path.to_path_buf(),
+                if inspect.authorization_args.authorize_with_raw_snapshot_file {
+                    BiscuitFormat::RawBiscuit
+                } else {
+                    BiscuitFormat::Base64Biscuit
+                },
+            )))
+        }
         // the other combinations are prevented by clap
         _ => unreachable!(),
     };
 
-    if let Some(vf) = &authorizer_from {
-        ensure_no_input_conflict(vf, &biscuit_from)?;
+    if let Some(AuthorizerInput::FromDatalog(dlf, _)) = &authorizer_from {
+        ensure_no_input_conflict(dlf, &biscuit_from)?;
     }
 
     if inspect.query_args.query.is_some() && public_key_from.is_none() {
@@ -390,8 +475,7 @@ pub fn handle_inspect_inner(inspect: &Inspect) -> Result<InspectionResults> {
         let external_key = external_keys
             .get(i)
             .expect("Incorrect block index")
-            .clone()
-            .map(hex::encode);
+            .map(|pk| pk.to_string());
         blocks.push(TokenBlock {
             code: biscuit.print_block_source(i)?,
             external_key,
@@ -399,6 +483,7 @@ pub fn handle_inspect_inner(inspect: &Inspect) -> Result<InspectionResults> {
                 .get(i)
                 .map(hex::encode)
                 .unwrap_or_else(|| "n/a".to_owned()),
+            version: biscuit.block_version(i)?,
         });
     }
 
@@ -413,36 +498,43 @@ pub fn handle_inspect_inner(inspect: &Inspect) -> Result<InspectionResults> {
     let query_result;
 
     if let Some(key_from) = public_key_from {
-        let key = read_public_key_from(&key_from)?;
+        let key = read_public_key_from(&key_from, &inspect.key_algorithm)?;
         let sig_result = biscuit.verify(key);
         signatures_check = Some(sig_result.is_ok());
 
         if let Ok(biscuit) = sig_result {
-            let mut authorizer_builder = biscuit.authorizer()?;
+            let mut authorizer_builder;
             if let Some(auth_from) = authorizer_from {
-                read_authorizer_from(
-                    &auth_from,
-                    &inspect.param_arg.param,
-                    &mut authorizer_builder,
-                )?;
+                authorizer_builder = read_authorizer_from(&auth_from)?;
                 if inspect.authorization_args.include_time {
                     let now = Utc::now().to_rfc3339();
                     let time_fact = format!("time({})", now);
-                    authorizer_builder.add_fact(time_fact.as_ref())?;
+                    authorizer_builder = authorizer_builder.fact(time_fact.as_ref())?;
                 }
-                let (_, _, _, policies) = authorizer_builder.dump();
+                if let Some(policies_snapshot_file) = &inspect.dump_policies_snapshot_to {
+                    if inspect.dump_raw_raw_policies_snapshot {
+                        let bytes = authorizer_builder.to_raw_snapshot()?;
+                        fs::write(policies_snapshot_file, bytes)?;
+                    } else {
+                        let str = authorizer_builder.to_base64_snapshot()?;
+                        fs::write(policies_snapshot_file, str)?;
+                    }
+                }
 
-                let authorizer_result = authorizer_builder.authorize_with_limits(RunLimits {
+                let mut authorizer = authorizer_builder.build(&biscuit)?;
+                let (_, _, _, policies) = authorizer.dump();
+
+                let authorizer_result = authorizer.authorize_with_limits(RunLimits {
                     max_facts: inspect
-                        .authorization_args
+                        .run_limits_args
                         .max_facts
                         .unwrap_or_else(|| RunLimits::default().max_facts),
                     max_iterations: inspect
-                        .authorization_args
+                        .run_limits_args
                         .max_iterations
                         .unwrap_or_else(|| RunLimits::default().max_iterations),
                     max_time: inspect
-                        .authorization_args
+                        .run_limits_args
                         .max_time
                         .map_or_else(|| RunLimits::default().max_time, |d| d.to_std().unwrap()),
                 });
@@ -457,16 +549,16 @@ pub fn handle_inspect_inner(inspect: &Inspect) -> Result<InspectionResults> {
                             )
                         })
                         .into(),
-                    iterations: authorizer_builder.iterations(),
-                    elapsed: authorizer_builder.execution_time(),
+                    iterations: authorizer.iterations(),
+                    elapsed_micros: authorizer.execution_time().map(|e| e.as_micros()),
                 });
 
                 if let Some(snapshot_file) = &inspect.dump_snapshot_to {
                     if inspect.dump_raw_snapshot {
-                        let bytes = authorizer_builder.to_raw_snapshot()?;
+                        let bytes = authorizer.to_raw_snapshot()?;
                         fs::write(snapshot_file, bytes)?;
                     } else {
-                        let str = authorizer_builder.to_base64_snapshot()?;
+                        let str = authorizer.to_base64_snapshot()?;
                         fs::write(snapshot_file, str)?;
                     }
                 }
@@ -476,19 +568,20 @@ pub fn handle_inspect_inner(inspect: &Inspect) -> Result<InspectionResults> {
                         query,
                         inspect.query_args.query_all,
                         &inspect.param_arg.param,
-                        &mut authorizer_builder,
+                        &mut authorizer,
                     )?);
                 } else {
                     query_result = None;
                 }
             } else {
                 auth_result = None;
+                let mut authorizer = biscuit.authorizer()?;
                 if let Some(query) = &inspect.query_args.query {
                     query_result = Some(handle_query(
                         query,
                         inspect.query_args.query_all,
                         &inspect.param_arg.param,
-                        &mut authorizer_builder,
+                        &mut authorizer,
                     )?);
                 } else {
                     query_result = None;
@@ -550,92 +643,83 @@ pub fn handle_inspect_snapshot_inner(
         BiscuitBytes::FromFile(snapshot_format, inspect_snapshot.snapshot_file.clone())
     };
 
-    let authorizer_from = match (
-        &inspect_snapshot.authorization_args.authorize_interactive,
-        &inspect_snapshot.authorization_args.authorize_with,
-        &inspect_snapshot.authorization_args.authorize_with_file,
-    ) {
-        (false, None, None) => None,
-        (true, None, None) => Some(DatalogInput::FromEditor),
-        (false, Some(str), None) => Some(DatalogInput::DatalogString(str.to_owned())),
-        (false, None, Some(path)) => Some(DatalogInput::FromFile(path.to_path_buf())),
-        // the other combinations are prevented by clap
-        _ => unreachable!(),
-    };
-
-    if let Some(vf) = &authorizer_from {
-        ensure_no_input_conflict(vf, &snapshot_from)?;
-    }
-
     let mut authorizer = read_snapshot_from(&snapshot_from)?;
-    let snapshot_description = SnapshotDescription {
+
+    let contents = SnapshotDescription {
         code: authorizer.to_string(),
         iterations: authorizer.iterations(),
-        elapsed_micros: authorizer.execution_time().as_micros(),
+        elapsed_micros: authorizer.execution_time().map(|d| d.as_micros()),
+        already_evaluated: authorizer.execution_time().is_some(),
     };
 
-    let auth_result;
-    let query_result;
-
-    if let Some(auth_from) = authorizer_from {
-        read_authorizer_from(
-            &auth_from,
-            &inspect_snapshot.param_arg.param,
-            &mut authorizer,
-        )?;
-        if inspect_snapshot.authorization_args.include_time {
-            let now = Utc::now().to_rfc3339();
-            let time_fact = format!("time({})", now);
-            authorizer.add_fact(time_fact.as_ref())?;
-        }
-        let (_, _, _, policies) = authorizer.dump();
-
-        let authorizer_result = authorizer.authorize_with_limits(RunLimits {
-            max_facts: inspect_snapshot
-                .authorization_args
-                .max_facts
-                .unwrap_or_else(|| RunLimits::default().max_facts),
-            max_iterations: inspect_snapshot
-                .authorization_args
-                .max_iterations
-                .unwrap_or_else(|| RunLimits::default().max_iterations),
-            max_time: inspect_snapshot
-                .authorization_args
-                .max_time
-                .map_or_else(|| RunLimits::default().max_time, |d| d.to_std().unwrap()),
-        });
-        auth_result = Some(AuthResult {
-            policies: policies.iter().map(|p| p.to_string()).collect::<Vec<_>>(),
-            result: authorizer_result
-                .map(|i| {
-                    (
-                        i,
-                        policies.get(i).expect("Incorrect policy index").to_string(),
-                    )
-                })
-                .into(),
-            iterations: authorizer.iterations(),
-            elapsed: authorizer.execution_time(),
-        });
-    } else {
-        auth_result = None;
-    }
-
-    if let Some(query) = &inspect_snapshot.query_args.query {
-        query_result = Some(handle_query(
-            query,
-            inspect_snapshot.query_args.query_all,
-            &inspect_snapshot.param_arg.param,
-            &mut authorizer,
-        )?);
-    } else {
-        query_result = None;
-    }
+    let evaluation = handle_snapshot_evaluation(inspect_snapshot, &mut authorizer);
 
     Ok(SnapshotInspectionResults {
-        snapshot: snapshot_description,
-        auth: auth_result,
-        query: query_result,
+        contents,
+        evaluation: evaluation.into(),
+    })
+}
+
+fn handle_snapshot_evaluation(
+    inspect_snapshot: &InspectSnapshot,
+    authorizer: &mut Authorizer,
+) -> std::result::Result<SnapshotEvaluationResults, Token> {
+    let _ = authorizer.run()?;
+
+    let (_, _, _, policies) = authorizer.dump();
+    let policies: Vec<String> = policies.into_iter().map(|p| p.to_string()).collect();
+
+    let authorizer_result = authorizer.authorize_with_limits(RunLimits {
+        max_facts: inspect_snapshot
+            .run_limits_args
+            .max_facts
+            .unwrap_or_else(|| RunLimits::default().max_facts),
+        max_iterations: inspect_snapshot
+            .run_limits_args
+            .max_iterations
+            .unwrap_or_else(|| RunLimits::default().max_iterations),
+        max_time: inspect_snapshot
+            .run_limits_args
+            .max_time
+            .map_or_else(|| RunLimits::default().max_time, |d| d.to_std().unwrap()),
+    });
+
+    let result = match authorizer_result {
+        Ok(policy_id) => RResult::Ok((
+            policy_id,
+            policies
+                .get(policy_id)
+                .expect("Incorrect policy index")
+                .to_string(),
+        )),
+        Err(e) => RResult::Err { error: e },
+    };
+
+    let auth = AuthResult {
+        policies,
+        result,
+        iterations: authorizer.iterations(),
+        elapsed_micros: authorizer.execution_time().map(|e| e.as_micros()),
+    };
+
+    let query = inspect_snapshot
+        .query_args
+        .query
+        .as_ref()
+        .map(|query| {
+            handle_query(
+                query,
+                inspect_snapshot.query_args.query_all,
+                &inspect_snapshot.param_arg.param,
+                authorizer,
+            )
+        })
+        .transpose()?;
+
+    Ok(SnapshotEvaluationResults {
+        iterations: authorizer.iterations(),
+        auth,
+        query,
     })
 }